webapp/requirements.txt
click==8.0.3
colorama==0.4.4
easyocr==1.4.1
Flask==2.0.2
imageio==2.9.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==8.2.0
pyparsing==3.0.6
python-bidi==0.4.2
PyWavelets==1.2.0
PyYAML==6.0
scikit-image==0.19.0
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
torch==1.10.0
torchvision==0.11.1
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
